Refactor: 온보딩 네비게이션
